SDR-297 refactor: 유저의 리뷰 조회 기능을 ReviewService로 이동
