+++ be/src/main/java/com/jjikmuk/sikdorak/review/service/ReviewService.java
@@ -28,12 +28,14 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ReviewService {
 
     private static final long FIRST_CURSOR_ID = 0L;
@@ -64,13 +66,12 @@ public class ReviewService {
 
     /*
     TODO - 개선사항
-     [] 비회원인 경우
-        [] public 상태의 리뷰만 제공한다.
-        [] 추천하는 피드들을 우선으로 준다. (좋아요 많은 순 / 평점이 높은 순 / 댓글 많은 순 등등)
+     [x] 비회원인 경우
+        [x] public 상태의 리뷰만 제공한다.
+        [x] 추천하는 피드들을 우선으로 준다. (좋아요 많은 순)
      [] 회원인 경우
-        [] public, protected 상태만 준다.
+        [x] public, protected 상태만 준다.
         [] 친구들의 피드들을 우선으로 준다. (최신순)
-        [] 유저 본인의 private 게시물도 있다면 함께 보여야 한다.
         [] 추천 조건을 변경해서 피드들을 따로 볼 수 있다. (비회원의 추천 피드와 동일하게)
      */
     @Transactional(readOnly = true)
@@ -87,6 +88,18 @@ public class ReviewService {
 
         return RecommendedReviewResponse.of(reviewsResponse, cursorPageResponse);
 
+    }
+
+    @Transactional(readOnly = true)
+    public List<ReviewDetailResponse> searchUserReviewsByUserIdAndRelationType(Long searchUserId, LoginUser loginUser) {
+        log.debug("searchByUserReviews: searchUserId={}, loginUser.id={}, loginUser.authority={}", searchUserId, loginUser.getId(), loginUser.getAuthority());
+
+        User searchUser = userRepository.findById(searchUserId)
+            .orElseThrow(NotFoundUserException::new);
+
+        List<Review> userReviews = findUserReviews(loginUser, searchUser);
+
+        return getReviewsResponse(userReviews);
     }
 
     @Transactional
@@ -264,5 +277,13 @@ public class ReviewService {
         return reviews.stream()
             .map(Review::getUserId)
             .toList();
+    }
+
+    private List<Review> findUserReviews(LoginUser loginUser, User searchUser) {
+        return switch (searchUser.relationTypeTo(loginUser)) {
+            case SELF -> reviewRepository.findByUserId(searchUser.getId());
+            case CONNECTION -> reviewRepository.findByUserIdAndConnection(searchUser.getId());
+            case DISCONNECTION -> reviewRepository.findByUserIdAndDisconnection(searchUser.getId());
+        };
     }
 }
